Currently broken: We need to sort things out with the rendering.
It may be better to step back and look at a properly-done example
that integrates ImGui with a Vulkan renderer, so we can get
everything straightened out as it needs to be.

We should also read the docs on the various parts and pieces to get
clear on what each thing is really doing.

TODO